refactor of editpage into a more flexible, modular editor
